Updated high scores
Updated the High Scores page to query the database and print the
results. There is a problem printing the query results from the db.php
file. For some reason the only way that I can get the results to print
is to do so in an iFrame element. This is not the correct method, and
needs to be addressed.