Updated pages to include navBar
Updated the site to include navBar.html rather than having repeated
code throughout the site. Now if you need to update the navBar do so in
navBar.html.

Note that in the header element of a page under the title element you
need to include an id=“title” and pageName=“name_of_page” attributes
for the header to formate correctly.