Added an php script for adding scores
Added a php script for adding scores after a user plays a game. Also
updated the database to include a friends table. Note that this table
will allow adding duplicate friends. So we will need to do a check to
make sure that a user is not already friends with a person.